Fetch component update fix
Allowing for the user to fetch components based on the handle of the component, fixing readme for submodules to accurately state what has been changed

+++ Mage/Source/ECS/mageComponent.c
@@ -1,6 +1,6 @@
 #include <mageAPI.h>
 
-static void mageDefaultConstructer(void *data, const uint32_t size)
+static void mageDefaultConstructer(void *data, const uint64_t size)
 {
 
 }
